Added title and date fields to quote format.
Without a "title" field there is no way to mark up quote credits with
the cite tag in a way that conforms to the HTML5 spec. Cite is only
meant to be applied to the title of a work, not the name of a person. I
included date since a year is often given. Ultimately this will allow
theme authors to follow the "Name, Title, Date" convention of quote
credits while conforming to the HTML5 spec.