Added shortcode_atts_gallery filter that allows us to intercept the gallery short code and maipulate it.
Need to work out how to properly parse out the shortcode value that a
user enters and add the id attributes to the regular gallery tag.

What if the user puts extra atts in the shortcode? Or do we only allow
ids and maybe they add other atts to the tag in thd body?